+++ src/app/bom/bom-viewer/bom-viewer.component.ts
@@ -1,142 +1,21 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { TreeModel } from 'ng2-tree';
-import { ApiDataBomService } from '../../_shared/services/api-data/_index';
-import { Subscription } from 'rxjs/Subscription';
-import { ViewEncapsulation } from '@angular/core';
-
-declare var Treant: any;
-declare var $: any;
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-  // encapsulation: ViewEncapsulation.None,
   selector: 'app-bom-viewer',
   templateUrl: './bom-viewer.component.html',
   styleUrls: ['./bom-viewer.component.css', '../../_shared/styles/common.css']
 })
 export class BomViewerComponent implements OnInit {
 
-  @ViewChild('treeComponent') treeComponent;
+  selectedBom: any;
 
-  billList: any;  // for getting the list of bills in drop-down
-  billListSub: Subscription;
-  bill: any;  // for storing the selected bill as flat array
-
-  bomChart: any;  // treant chart object
-  bomChartOptions: any; // object of options for treant chart
-
-
-  constructor(private apiDataBomService: ApiDataBomService) { }
+  constructor() { }
 
   ngOnInit() {
-    // get list of bills in drop-down
-    this.billListSub = this.apiDataBomService.index().subscribe( res => {
-      this.billList = res;
-    });
-
   }
 
-  onBomSelect(selected: number) {
-    this.bomChartOptions = {};
-
-    // get the selected BOM as flat array
-    const bomSubscription = this.apiDataBomService.showSingleBom(selected).subscribe( res => {
-
-
-      this.bill = res;
-      bomSubscription.unsubscribe();
-      // console.log(this.bill);
-
-      this.bomChartOptions = {
-        chart: {
-          container: '#tree-simple',
-          levelSeparation: 100,
-          siblingSeparation: 20,
-          // padding: 100,
-          connectors: {type: 'curve'},
-          rootOrientation: 'NORTH',
-          nodeAlign: 'TOP',
-          node: {collapsable: true}
-        }
-      };
-      // initialize bomtree
-      this.bomChartOptions.nodeStructure = {
-        text: {
-          name: {val: this.bill[0].ParentName},
-          title: `${this.bill[0].ParentEntity} - ${this.bill[0].ParentType}`
-        },
-        id: this.bill[0].ParentID
-      };
-
-      // recursively parse the BOM structure
-      const jsonBom = this.bomTraverse(0, 1);
-
-      // add the recursive output as 'children' property of the tree nodeStructure
-      this.bomChartOptions.nodeStructure.children = jsonBom.nextLvData;
-
-      // if the parent has more than 9 children, initailize it collapsed
-      if (this.bomChartOptions.nodeStructure.children.length > 9) {
-        this.bomChartOptions.nodeStructure.collapsed = true;
-      }
-
-      if (this.bomChart) {
-        this.bomChart.destroy();
-      }
-      this.bomChart = new Treant(this.bomChartOptions, this.onTreeLoadComplete, $);
-      console.log('finalized bom structure');
-      console.log(this.bomChartOptions.nodeStructure);
-    });
+  onBomSelect(selection: any) {
+    this.selectedBom = selection;
   }
-
-  onTreeLoadComplete() {
-    // callback function executed when Treant is done drawing the tree, currently unused
-    // console.log('completed');
-  }
-
-  bomTraverse(i: number, lv: number) {
-    // i = index of the array to start traversing (usu 0)
-    // lv - initial level of the BOM (usu 1)
-    const children = [];
-    while (i < this.bill.length) {
-      if (this.bill[i].Level === lv) {
-        // traverse down and collect all the siblings in this level
-        let newNode: any;
-        newNode = {
-          text: {
-            name: this.bill[i].ChildName,
-            title: `${this.bill[i].ChildType} | Qty: ${this.bill[i].QtyPer}`,
-            // desc: `Qty: ${this.bill[i].QtyPer}`
-          },
-          id: this.bill[i].ChildID
-        };
-        if (this.bill[i].ChildEntity === 'Part') {
-          newNode.HTMLclass = this.bill[i].ChildDepartment;
-        }
-        children.push(newNode);
-        i++;
-      } else if (this.bill[i].Level > lv) {
-        // if the next record is a child, recurse
-        // when we return to this level, continue traversing from the farthest-reached index
-        const output = this.bomTraverse(i, lv + 1);
-        const lastIndex = children.length - 1;
-        children[lastIndex].children = output.nextLvData;
-        // if there are 10+ children, initialize the node as collapsed
-        if (output.nextLvData.length > 9) {
-          children[lastIndex].collapsed = true;
-        }
-        i = Number(output.nextRow);
-      } else if (this.bill[i].Level < lv) {
-        // if the next record is a parent, return the complete set of nested children
-        // and the next value to continue traversing at
-        return {
-          nextRow: i,
-          nextLvData: children
-        };
-      }
-    } // end while
-    return {
-      nextRow: i,
-      nextLvData: children
-    };
-  } // end bomTraverse
 
 }